tsx + useListProxy test
